+++ modules/apps/oauth2-provider/oauth2-provider-scope-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay OAuth2 Provider Scope API
+Bundle-SymbolicName: com.liferay.oauth2.provider.scope.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.oauth2.provider.scope
