+++ modules/apps/oauth2-provider/oauth2-provider-scope-api/build.gradle
