+++ modules/apps/oauth2-provider/oauth2-provider-scope-api/src/main/java/com/liferay/oauth2/provider/scope/ScopeChecker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.scopes;
+package com.liferay.oauth2.provider.scope;
 
 import aQute.bnd.annotation.ProviderType;
