+++ modules/apps/oauth2-provider/oauth2-provider-scope-api/src/main/resources/com/liferay/oauth2/provider/scope/packageinfo
